requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
brotlipy==0.7.0
    # via -r requirements.in
certifi==2022.5.18.1
    # via sentry-sdk
cffi==1.15.0
    # via brotlipy
colorama==0.4.4
    # via ipython
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.13
    # via
    #   -r requirements.in
    #   django-filter
    #   django-import-export
    #   django-js-reverse
    #   django-log-request-id
    #   django-model-utils
    #   djangorestframework
django-debreach==2.1.0
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-import-export==2.8.0
    # via -r requirements.in
django-js-reverse==0.9.1
    # via -r requirements.in
django-log-request-id==2.0.0
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-webpack-loader==1.5.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
executing==0.8.3
    # via stack-data
gunicorn==20.1.0
    # via -r requirements.in
ipython==8.4.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.3
    # via ipython
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
parso==0.8.3
    # via jedi
pickleshare==0.7.5
    # via ipython
pillow==9.1.1
    # via -r requirements.in
plotly==5.9.0
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via ipython
psutil==5.9.1
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via ipython
python-decouple==3.6
    # via -r requirements.in
pytz==2022.1
    # via
    #   django
    #   djangorestframework
pyyaml==6.0
    # via tablib
sentry-sdk==1.5.12
    # via -r requirements.in
six==1.16.0
    # via asttokens
sqlparse==0.4.2
    # via django
stack-data==0.2.0
    # via ipython
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
tenacity==8.0.1
    # via plotly
traitlets==5.2.1.post0
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.9
    # via sentry-sdk
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==6.1.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
